un-inline PrecludeAction
