+++ src/main/resources/fishingResources/localization/eng/UIStrings.json
@@ -4,6 +4,11 @@
       "Choose a card to put on top of your draw pile and make free until played."
     ]
   },
+  "${ModID}:PrecludeAction": {
+    "TEXT": [
+      "~Cheater...~ NL @CHEEEATER!@"
+    ]
+  },
   "${ModID}:FishDescriptor": {
     "TEXT": [
       "Fish"
