+++ src/main/resources/fishingResources/localization/kor/UIStrings.json
@@ -4,6 +4,11 @@
       "뽑을 카드 더미 맨 위로 올려놓을 카드를 선택하세요. 그 카드의 비용은 사용하기 전까지 0이 됩니다."
     ]
   },
+  "fishing:PrecludeAction": {
+    "TEXT": [
+      "~비겁한...~ NL @이@ 비겁한@ @놈!@"
+    ]
+  },
   "fishing:FishDescriptor": {
     "TEXT": [
       "물고기"
